Iterative/recursive procedure works; now need to make it efficient with a priority queue or perhaps multiple (need to define priority)

+++ hashcat/gengraph.c
@@ -47,7 +47,7 @@ int main(int argc, const char** argv) {
     char *rule_line = NULL;
     size_t read_amount = 0;
     while(getline(&rule_line, &read_amount, fp_rules) != -1) {
-        if(strlen(rule_line) > 0 && rule_line[0] != '#' && rule_line[0] != '\n') {
+        if(strlen(rule_line) > 0 && rule_line[0] != '#' && rule_line[0] != '\n' && rule_line[0] != ':') {
             rule_count++;
         }
         free(rule_line);
@@ -60,7 +60,7 @@ int main(int argc, const char** argv) {
     int rule_num = 0;
     int rule_len;
     while(getline(&rule_line, &read_amount, fp_rules) != -1) {
-        if(strlen(rule_line) > 0 && rule_line[0] != '#' && rule_line[0] != '\n') {
+        if(strlen(rule_line) > 0 && rule_line[0] != '#' && rule_line[0] != '\n' && rule_line[0] != ':') {
             rule_line[strlen(rule_line)-1] = 0; // kill delimiter
             rule_len = strlen(rule_line);
             rules[rule_num] = (char*)malloc(rule_len+1);
@@ -82,28 +82,14 @@ int main(int argc, const char** argv) {
         return EXIT_FAILURE;
     }
 
+    // clear db
     neo4j_result_stream_t *results;
-    
-    /*
-    neo4j_result_t *result = neo4j_fetch_next(results);
-    if (result == NULL)
+    results = neo4j_run(neo4j_con, "MATCH (n) DETACH DELETE n", neo4j_null);
+    if (results == NULL)
     {
-        neo4j_perror(stderr, errno, "Failed to fetch result");
-        return EXIT_FAILURE;
-    }
-
-    while(1) {
-        neo4j_result_t *result = neo4j_fetch_next(results);
-        if (result == NULL)
-        {
-            break;
-        }
-        neo4j_value_t value = neo4j_result_field(result, 0);
-        char buf[128];
-        printf("%s\n", neo4j_tostring(value, buf, sizeof(buf)));
+        neo4j_perror(stderr, errno, "Failed to run DELETE statement");
     }
     neo4j_close_results(results);
-    */
 
     char query[1024];
     char *password = NULL;
@@ -111,52 +97,76 @@ int main(int argc, const char** argv) {
     while(getline(&password, &line_restrict, fp_passwords) != -1) {
         memset(query, 0, 1024);
         password[strlen(password)-1] = 0; // cut off delimiter
-        printf("password: %s\n", password);
+        printf("adding password: %s\n", password);
         char *md5pass_hex = md5(password);
-        printf("password md5: %s\n", md5pass_hex);
+        printf("adding password md5: %s\n", md5pass_hex);
 
-        sprintf(query, "merge (p:Password { md5: '%s', password: '%s'})", md5pass_hex, password);
+        sprintf(query, "merge (p:Password { md5: '%s', password: '%s', iteration: 0})", md5pass_hex, password);
         results = neo4j_run(neo4j_con, query, neo4j_null);
         if (results == NULL)
         {
             neo4j_perror(stderr, errno, "Failed to run CREATE statement");
         }
         neo4j_close_results(results);
-
-        for(int rule_idx = 0; rule_idx < rule_count; rule_idx++) {
-            printf("rule: %s\n", rules[rule_idx]);
-            char password_result[RP_PASSWORD_SIZE];
-            memset(password_result, 0, sizeof(password_result));
-            _old_apply_rule(rules[rule_idx], strlen(rules[rule_idx]), password, (u32)strlen(password), password_result);
-            printf("result: %s\n", password_result);
-            char *md5result_hex = md5(password_result);
-            printf("result md5: %s\n", md5result_hex);
-
-            sprintf(query, "merge (p:Password { md5: '%s', password: '%s'})", md5result_hex, password_result);
-            results = neo4j_run(neo4j_con, query, neo4j_null);
-            if (results == NULL)
-            {
-                neo4j_perror(stderr, errno, "Failed to run CREATE statement");
-            }
-            neo4j_close_results(results);
-
-            sprintf(query, "match (before:Password), (after:Password) where before.md5='%s' and after.md5='%s' create (before)-[:RULE{ rule:'%s' }]->(after)",
-                    md5pass_hex, md5result_hex, rules[rule_idx]);
-            results = neo4j_run(neo4j_con, query, neo4j_null);
-            if (results == NULL)
-            {
-                neo4j_perror(stderr, errno, "Failed to run CREATE statement");
-            }
-            neo4j_close_results(results);
-
-            free(md5result_hex);
-        }
-        free(md5pass_hex);
         free(password);
+        free(md5pass_hex);
         password = NULL;
     }
     free(password);
     fclose(fp_passwords);
+
+    // loop per password found in neo4j
+    for(int iter = 0; iter < 10; iter++) {
+        printf("\n\n== Iteration %d\n\n", iter);
+        sprintf(query, "match (p:Password) where p.iteration = %d return p.password", iter);
+        results = neo4j_run(neo4j_con, query, neo4j_null);
+        while(1) {
+            neo4j_result_t *result = neo4j_fetch_next(results);
+            if(result == NULL)
+            {
+                break;
+            }
+            neo4j_value_t value = neo4j_result_field(result, 0);
+            char password[RP_PASSWORD_SIZE];
+            printf("password: %s\n", neo4j_string_value(value, password, RP_PASSWORD_SIZE));
+            neo4j_string_value(value, password, RP_PASSWORD_SIZE);
+            char *md5pass_hex = md5(password);
+            //printf("password md5: %s\n", md5pass_hex);
+
+            neo4j_result_stream_t *sub_results;
+            for(int rule_idx = 0; rule_idx < rule_count; rule_idx++) {
+                //printf("  rule: %s\n", rules[rule_idx]);
+                char password_result[RP_PASSWORD_SIZE];
+                memset(password_result, 0, sizeof(password_result));
+                _old_apply_rule(rules[rule_idx], strlen(rules[rule_idx]), password, (u32)strlen(password), password_result);
+                //printf("  result: %s\n", password_result);
+                char *md5result_hex = md5(password_result);
+                //printf("  result md5: %s\n", md5result_hex);
+
+                sprintf(query, "merge (p:Password { md5: '%s', password: '%s'}) on create set p.iteration = %d", md5result_hex, password_result, iter+1);
+                sub_results = neo4j_run(neo4j_con, query, neo4j_null);
+                if (results == NULL)
+                {
+                    fprintf(stderr, "%s", query);
+                    neo4j_perror(stderr, errno, "Failed to run CREATE statement");
+                }
+                neo4j_close_results(sub_results);
+
+                sprintf(query, "match (before:Password), (after:Password) where before.md5='%s' and after.md5='%s' create (before)-[:RULE{ rule:'%s' }]->(after)",
+                        md5pass_hex, md5result_hex, rules[rule_idx]);
+                sub_results = neo4j_run(neo4j_con, query, neo4j_null);
+                if (results == NULL)
+                {
+                    fprintf(stderr, "%s", query);
+                    neo4j_perror(stderr, errno, "Failed to run CREATE statement");
+                }
+                neo4j_close_results(sub_results);
+
+                free(md5result_hex);
+            }
+        }
+        neo4j_close_results(results);
+    }
 
     for(int rule_idx = 0; rule_idx < rule_count; rule_idx++) {
         free(rules[rule_idx]);
